requirements.txt
pandas
plotnine
duckdb==1.3.2
ibis-framework[duckdb]
nbval
# Pinned to ensure pyarrow is available as a prebuilt wheel to avoid local builds
pyarrow==12.0.1
# Pin a recent datasets release; change if you prefer another version
datasets==2.14.2
# Pin numpy version to avoid _ARRAY_API errors and other compatability issues
numpy==1.26.4
